pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scaling-operator-learning"
version = "0.1.0"
description = "Scaling laws for operator learning: resolution as a first-class scaling axis."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Lucas Perrier"}]
dependencies = [
  "numpy>=1.24",
  "scipy>=1.10",
  "torch>=2.1",
  "pyyaml>=6.0",
  "tqdm>=4.66",
  "pandas>=2.0",
  "matplotlib>=3.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.3",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B"]
